Added Source Folder to Repository Structure
Added to the file structure of the repository. Created a source folder to use for holding java packages. Created the structure for the java pacakge following standard java conventions.

Placed placeholder text files in the client and tools package folders with simple directions regarding the repository structure. Added example tool subpackage "Stamp" with basic README file describing the purpose of the tool.