Update to ToolInterface and implementation of abstract Tool class.
Changed name of interface from Tool to ToolInterface so that an abstract Tool class could be implemented. Tools can inherit from this class to get default implementation of Tool Interface.

This makes it so that the developer does not have to implement empty classes for actions their class does not handle.