Fixed Package Structure Integrity. Modifications to Tool Button. Added Deselect Handling.
CopyPaste needed to be placed in its own subfolder to ensure the integrity of the package structure. As such, the package has been modified for this tool.

Corrected an override issue with the toolbar button that caused program to crash whenever tool was deselected. Added handling for changing button background color to indicate its state.

Added a call to the keypresshandler during deselect since this contains functionality for "undoing" the selection bounding box drawn.

Minor logic modification to fix bounding box drawing bug.

Signed-off-by: Jeremy Maxey-Vesperman <maxe2470@kettering.edu>